check if target mounted before copying log ;)

+++ lib/util-menu.sh
@@ -47,10 +47,12 @@ main_menu_online() {
         *) dialog --backtitle "$VERSION - $SYSTEM ($ARCHI)" --yesno "$_CloseInstBody" 0 0
             if [[ $? -eq 0 ]]; then
                 check_for_error "exit installer."
-                dialog --backtitle "$VERSION - $SYSTEM ($ARCHI)" --yesno "\n$_LogInfo\n" 0 0
-                if [[ $? -eq 0 ]]; then
-                  [[ -e /mnt/.m-a.log ]] && cat ${LOGFILE} >> /mnt/.m-a.log
-                  install -m700 ${LOGFILE} /mnt/.m-a.log
+                if [[ -e /mnt/usr ]]; then
+                  dialog --backtitle "$VERSION - $SYSTEM ($ARCHI)" --yesno "\n$_LogInfo\n" 0 0
+                  if [[ $? -eq 0 ]]; then
+                    [[ -e /mnt/.m-a.log ]] && cat ${LOGFILE} >> /mnt/.m-a.log
+                    install -m700 ${LOGFILE} /mnt/.m-a.log
+                  fi
                 fi
                 umount_partitions
                 clear
